separate components for info and footer
