+++ src/components/Info.jsx
@@ -1,0 +1,8 @@
+const Info = () => (
+  <div>
+    Check how much a Solana wallet has spent on transaction fees, by entering
+    it&apos;s address. ☝️
+  </div>
+);
+
+export default Info;
